Improve comment section UI

+++ portfolio/src/main/java/com/google/sps/dao/ICommentDao.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface ICommentDao {
     List<Comment> getAllComments();
 
-    Comment getCommentById(int id);
+    Comment getCommentById(long id);
 
     ValidationResponse addComment(Comment comment);
 }
